Migrate fjs/effects/list to .f.mjs, split type-level API into types.ts
No proof.f.ts existed for this module (AGENTS.md's coverage gap is
pre-existing, not introduced here). Updates 8 dependents across cas,
mcp, media/type, and effects/node to import runtime values from
.f.mjs and types from types.ts separately.
