+++ fjs/cas/evo/module.f.ts
@@ -53,7 +53,7 @@ import { tryUtf8 } from '../../text/module.f.mjs'
 import { decodeText, encodeText, dialect, checkReferences, isHash, type LockMap, type Revision } from '../../media/revision/module.f.ts'
 import type { Ok, Result } from '../../types/result/types.ts'
 import { ok, error } from '../../types/result/module.f.mjs'
-import { nonEmpty, empty as elEmpty } from '../../effects/list/module.f.ts'
+import { nonEmpty, empty as elEmpty } from '../../effects/list/module.f.mjs'
 import { at, definedEntries } from '../../types/object/module.f.mjs'
 import type { StringMap } from '../../types/object/types.ts'
 import { unwrap } from '../../types/nullable/module.f.mjs'
