+++ fjs/cas/evo/proof.f.ts
@@ -9,7 +9,7 @@ import { cBase32ToVec, vecToCBase32 } from '../../basen/cbase32/module.f.mjs'
 import { unwrap } from '../../types/nullable/module.f.mjs'
 import type { Ok } from '../../types/result/types.ts'
 import { ok, error } from '../../types/result/module.f.mjs'
-import { nonEmpty, empty as elEmpty } from '../../effects/list/module.f.ts'
+import { nonEmpty, empty as elEmpty } from '../../effects/list/module.f.mjs'
 import type { IoResult } from '../../effects/node/module.f.ts'
 import { tryUtf8 } from '../../text/module.f.mjs'
 import { dialect as revisionDialect } from '../../media/revision/module.f.ts'
