+++ fjs/cas/proof.f.ts
@@ -10,7 +10,8 @@ import type { Ok } from '../types/result/types.ts'
 import { error, ok } from '../types/result/module.f.mjs'
 import { emptyState, virtual } from '../effects/node/virtual/module.f.ts'
 import { join } from '../path/module.f.mjs'
-import { nonEmpty, empty, type List } from '../effects/list/module.f.ts'
+import { nonEmpty, empty } from '../effects/list/module.f.mjs'
+import type { List } from '../effects/list/types.ts'
 import { assert, assertEq, assertNotNullish } from '../asserts/module.f.mjs'
 
 const testDir = './test-cas-cli'
