+++ fjs/effects/node/module.f.ts
@@ -22,7 +22,7 @@ import { ok, error as resultError, mapOk } from '../../types/result/module.f.mjs
 import type { StringMap } from '../../types/object/types.ts'
 import { do_, mapStep, okStep, pure, step } from '../module.f.mjs'
 import type { Effect, Func, Operation, ToAsyncOperationMap } from '../types.ts'
-import type { List } from '../list/module.f.ts'
+import type { List } from '../list/types.ts'
 
 export type IoResult<T> = Result<T, unknown>
