+++ fjs/effects/todo/fold-stream-combinator.md
@@ -53,13 +53,13 @@ in sync.
 
 ### Proposal
 
-Add a `foldStream` combinator to `fjs/effects/list/module.f.ts` whose step
+Add a `foldStream` combinator to `fjs/effects/list/module.f.mjs` whose step
 returns an `Effect`, so it subsumes both the pure folds (step = `pure(...)`)
 and the effectful writers.
 
 **Layering note:** `IoResult` cannot appear in this module's signature —
 it is exported by `fjs/effects/node/module.f.ts`, which already imports
-`List` from `fjs/effects/list/module.f.ts`, so importing it back would be
+`List` from `fjs/effects/list/module.f.mjs`, so importing it back would be
 a cycle. But `IoResult<T>` is just an alias for `Result<T, unknown>` from
 `fjs/types/result` (a types-layer module `effects/list` can import freely),
 so write the signature in terms of `Result` — call sites that hold
@@ -101,7 +101,7 @@ consumers first; the writers follow only if the shape stays clean.
 
 ### Tasks
 
-- [ ] Add `foldStream` to `fjs/effects/list/module.f.ts` with proof coverage.
+- [ ] Add `foldStream` to `fjs/effects/list/module.f.mjs` with proof coverage.
 - [ ] Convert `detectStream` (`fjs/media/type`) and `collectRead` (`fjs/cas`).
 - [ ] Convert `writeLoop` (`fjs/effects/node`) if the effectful step fits
       without contortion; otherwise document why in this issue and keep it.
