+++ fjs/media/type/module.f.ts
@@ -38,7 +38,7 @@ import { iterable } from '../../types/list/module.f.mjs'
 import type { Nullable } from '../../types/nullable/types.ts'
 import { pure, step } from '../../effects/module.f.mjs'
 import type { Effect, Operation } from '../../effects/types.ts'
-import type { List } from '../../effects/list/module.f.ts'
+import type { List } from '../../effects/list/types.ts'
 import type { IoResult } from '../../effects/node/module.f.ts'
 import { ok, error } from '../../types/result/module.f.mjs'
 import { isValidCodePoint, isTextCodePoint } from '../../text/code_point/module.f.mjs'
